+++ R/release.r
@@ -69,13 +69,10 @@ release <- function(pkg = ".", check = TRUE) {
     }
   }
 
-  if (new_pkg) {
-    policies <- paste("Have you read and do you agree to the the CRAN policies?",
-      "\n(http://cran.r-project.org/web/packages/policies.html)")
-
-    if (yesno(policies))
-      return(invisible())
-  }
+  policies <- paste("Have you read and do you agree to the the CRAN policies?",
+    "\n(http://cran.r-project.org/web/packages/policies.html)")
+  if (yesno(policies))
+    return(invisible())
 
   if (yesno("Have you checked on win-builder (with build_win())?"))
     return(invisible())
@@ -86,7 +83,9 @@ release <- function(pkg = ".", check = TRUE) {
       return(invisible())
   }
 
+  rule("DESCRIPTION")
   cat(readLines(file.path(pkg$path, "DESCRIPTION")), sep = "\n")
+  cat("\n")
   if (yesno("Is DESCRIPTION up-to-date?"))
     return(invisible())
 
@@ -107,22 +106,18 @@ release <- function(pkg = ".", check = TRUE) {
     }
   }
 
-  if (yesno("Ready to upload?"))
+  if (yesno("Is your email address ", maintainer(pkg)$email, "?"))
     return(invisible())
 
-  message("Building")
-  built_path <- build(pkg, tempdir())
-  message("File size: ", file.info(built_path)$size, " bytes")
+  rule("cran-comments.md ")
+  cat(cran_comments(pkg), "\n\n")
+  if (yesno("Are the CRAN submission comments correct?"))
+    return(invisible())
 
-  message("Uploading")
-  ftpUpload(built_path, paste("ftp://cran.R-project.org/incoming/",
-    basename(built_path), sep = ""))
+  if (yesno("Ready to submit?"))
+    return(invisible())
 
-  message("Preparing email")
-  body <- release_email(pkg$package, new_pkg)
-  subject <- paste("CRAN submission ", pkg$package, " ", pkg$version, sep = "")
-  email("cran@r-project.org", subject, body)
-  message("Remember to send the email as plain text ASCII and not HTML.")
+  submit_cran(pkg)
 
   if (file.exists(file.path(pkg$path, ".git"))) {
     message("Don't forget to tag the release when the package is accepted!")
@@ -148,11 +143,11 @@ release_email <- function(name, new_pkg) {
     sep = "")
 }
 
-yesno <- function(question) {
+yesno <- function(...) {
   yeses <- c("Yes", "Definitely", "For sure", "Yup", "Yeah")
   nos <- c("No way", "Not yet", "I forgot", "No", "Nope")
 
-  cat(question)
+  cat(paste0(..., collapse = ""))
   qs <- c(sample(yeses, 1), sample(nos, 2))
   rand <- sample(length(qs))
 
@@ -193,4 +188,104 @@ email_browser <- function() {
 
   browser <- Sys.which(c("xdg-open", "open"))
   browser[nchar(browser) > 0][[1]]
+}
+
+
+maintainer <- function(pkg = ".") {
+  pkg <- as.package(pkg)
+
+  authors <- pkg$`authors@r`
+  if (!is.null(authors)) {
+    people <- eval(parse(text = authors))
+    if (is.character(people)) {
+      maintainer <- as.person(people)
+    } else {
+      maintainer <- Find(function(x) "cre" %in% x$role, people)
+    }
+  } else {
+    maintainer <- pkg$maintainer
+    if (is.null(maintainer)) {
+      stop("No maintainer defined in package.", call. = FALSE)
+    }
+    maintainer <- as.person(maintainer)
+  }
+
+  list(
+    name = paste(maintainer$given, maintainer$family),
+    email = maintainer$email
+  )
+}
+
+cran_comments <- function(pkg = ".") {
+  pkg <- as.package(pkg)
+
+  path <- file.path(pkg$path, "cran-comments.md")
+  if (!file.exists(path)) {
+    stop("Can't find cran-comments.md in ", pkg$package, ".\n",
+      "This file for comments for package submission and must exist.\n",
+      "Please create and add to .Rbuildignore (with add_build_ignore())",
+      call. = FALSE)
+  }
+
+  paste0(readLines(path, warn = FALSE), collapse = "\n")
+}
+
+cran_submission_url <- "http://xmpalantir.wu.ac.at/cransubmit/index2.php"
+
+#' Submit a package to CRAN.
+#'
+#' This uses the new CRAN web-form submission process. After submission, you
+#' will recieve an email asking you to confirm submission - this is used
+#' to check that the package is submitted by the maintainer.
+#'
+#' It's recommend that you use \code{\link{release}()} rather than this
+#' function as it performs more checks prior to submission.
+#'
+#' @param pkg package description, can be path or package name.  See
+#'   \code{\link{as.package}} for more information
+#' @export
+#' @keywords internal
+submit_cran <- function(pkg = ".") {
+  pkg <- as.package(pkg)
+  maint <- maintainer(pkg)
+  comments <- cran_comments(pkg)
+
+  message("Building")
+  built_path <- build(pkg, tempdir())
+  message("File size: ", file.info(built_path)$size, " bytes")
+
+  # Initial upload ---------
+  message("Uploading package & comments")
+  body <- list(
+    pkg_id = "",
+    name = maint$name,
+    email = maint$email,
+    uploaded_file = httr::upload_file(built_path, "application/x-gzip"),
+    comment = comments,
+    upload = "Upload package"
+  )
+  r <- httr::POST(cran_submission_url, body = body)
+  stop_for_status(r)
+  new_url <- parse_url(r$url)
+  new_url$query$strErr
+
+  # Confirmation -----------
+  message("Confirming submission")
+  body <- list(
+    pkg_id = new_url$query$pkg_id,
+    name = maint$name,
+    email = maint$email,
+    policy_check = "1/",
+    submit = "Submit package"
+  )
+  r <- httr::POST(cran_submission_url, body = body)
+  new_url <- parse_url(r$url)
+  if (new_url$query$submit == "1") {
+    message("Package submission successful.\n",
+      "Check your email for confirmation link.")
+  } else {
+    stop("Package failed to upload.", call. = FALSE)
+  }
+
+  invisible(TRUE)
 }