+++ NEWS.md
@@ -1,3 +1,6 @@
+* `release()` uses new CRAN submission process, as implemented by 
+  `submit_cran()` (#430).
+
 * You can add arbitrary extra questions to `release()` by defining a function 
   `release_questions()` in your package. Your `release_questions()` should 
   return a character vector of questions to ask (#451).